Translate some more group code from Python validator and fix bug
I have added formatHedTag, which removes leading and trailing double
quotes and slashes, returning the formatted tag in lowercase.
I also created a stub for top-level tags, which will be implemented
once I implement proper group parsing in the splitter. A bug in the
splitter related to checking for invalid characters has been fixed.
Tests for the formatting code (which pass) and for the top-level
tag code (which fail for now) were added.