Revamp tag parsing
This moves the different parsed versions of HED tags within a new
object, which includes the original, formatted, and canonical (long)
forms, along with the bounds in the original string and any issues
found during parsing.

I still need to incorporate the bounds into the issue messages and
fix the tests. The event.js test all pass, and the HED 3 tests in
that file have been migrated to the local copy of the HED v8.0.0-alpha
schema.