Remove tilde validation
Tildes are being removed from HED 3, and tilde validation is being
removed retroactively for previous generations of HED. This commit
removes the tag group tilde count check, adds the tilde to the invalid
character list, removes tests including tildes, and adds a invalid
character test for tildes.

Also fixed in this commit was an issue with short-to-long tag conversion
issues not being propagated through the string parsing code.