Delete expected boolean validation result from dataset and event tests
The boolean results from the validator functions, whose expected values
were included in expectedResults, were redundant to expectedIssues,
as true values always corresponded to empty issue lists. Furthermore,
when strings that were expected to be valid failed, the expected "true"
boolean values were preempting the unexpected issues from displaying,
which meant I had to temporarily change the values to see the errors
in order to fix them.

expectedResults instances in other files were kept, as they are not
redundant to the issue list.