Add parser for new schema spec
Also rewrite special element handling in XPath implementation to be
more generic.

The HED 3 tests work unmodified with the alpha.3 schema (not included),
though some edits had to be made to the schema itself (namely adding
unitModifier). More tests will be added later, and the schema tests
will have to be rewritten.