Initial implementation of library schema nickname parsing
This commit adds the necessary code to parse library schema nicknames
from HED tags and pass them to ParsedHedTag objects. It also adds the
unmatchedLibrarySchema issue for cases when a schema cannot be found.

The internal converter functions have been modified to accept Schema
objects instead of Schemas objects, and the public-facing ones
(which were not changed for BC reasons) have been deprecated from the
public API and will be made private in version 4.0.0 due to a lack of
independent utility.

A minor issue in the new schema loading code was also fixed, bringing
it in line with the surrounding code, and some documentation was fixed.

None of this code has been tested with tags using library schemas. Tests
will be added in the next commit.