Refactor value validation
This commit moves the value validation checks to the generation-specific
classes. It also adds a utility property for the formatted takes-value
tag on a ParsedHedTag.

The tests currently fail due to an issue with Definition tags with
placeholders not generating a takesValueTag in the ParsedHed3Tag object,
thus leading to an undefined-pointer error. I don't have a good solution
at this time.